requirements.txt
taipy
openai
pandas
llama-index
pypdf 
pypdf2 
python-dotenv